RegistrationPageSamplecode
